Use custom wrapping legend in per-level expanded views
Apply the same non-overflowing legend fix from クリアランプ概要 to the four
per-level expanded views (clear type, clear-type trends, DJ level, DJ
level trends): hide the built-in chart legend and render the compact
two-column legend below. Generalize the legend into AnalyticsChartLegend
with ClearTypeLegend/DJLevelLegend specializations, and add
IIDXDJLevel.color for the DJ-level palette.
